requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.3
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-jwt-refresh-token==0.5
drf-yasg==1.17.1
idna==2.9
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.1
PyJWT==1.7.1
pyparsing==2.4.6
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
psycopg2==2.8.4
gunicorn==20.0.4
git+https://github.com/AlexanderUstinovv/body_mass_calculator.git
git+https://github.com/AlexanderUstinovv/medical_test.git